ensuring popup call inti again when required
